% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_create_boxplots.R
\name{aa_create_boxplots}
\alias{aa_create_boxplots}
\title{Assess the accuracy of a set of ground control points with boxplots}
\usage{
aa_create_boxplots(gcp_data, export = F, filename, output_path)
}
\arguments{
\item{gcp_data}{object of class \verb{spatial points} as generated by \verb{merge_gcps())}.}

\item{export}{logical of length 1. Should the result be exported to a png ? Defaults to FALSE.}

\item{filename}{character, optional, name of the resulting png file}

\item{output_path}{character, optional, path to folder where the file will be exported to}
}
\value{
aa_create_boxplots: ggplot2 object
}
\description{
This function takes the output of \code{merge_gcps()} and provides boxplots for the differences per gcp in X,Y,Z directions
per site
}
\examples{
\dontrun{
generate boxplots:
GCPs_2017_ref <- st_read("I:/misc/GCPs/2017-06/GCPs_ALS_2017_ref.gpkg")
aa_create_boxplots(gcp_data = GCPs_2017_ref, export = T, filename = "Difference_GCPs_ALS_2017",
                   output_path = "I:/misc/GCPs/2017-06")
calculate metrics:
metrics <- aa_metrics(GCPs_2017_ref, export = T, filename = "Metrics_ALS_2017", output_path = "I:/misc/GCPs/2017-06")
metrics
}
re-format data for better visualization:
generate the plot:
export the plot to disk:
}
