% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_metrics.R
\name{aa_metrics}
\alias{aa_metrics}
\title{Assess the accuracy of a set of ground control points with boxplots}
\usage{
aa_metrics(gcp_data, export = F, filename, output_path)
}
\arguments{
\item{gcp_data}{object of class \verb{spatial points} as generated by \verb{merge_gcps())}.}

\item{export}{logical of length 1. Should the result be exported to a csv file? Defaults to FALSE.}

\item{filename}{character, optional, name of the resulting csv file}

\item{output_path}{character, optional, path to folder where the csv file will be exported to}
}
\value{
data frame
}
\description{
This function takes the output of \code{merge_gcps()} and provides accuracy metrics for the differences per gcp in X,Y,Z directions
per site
}
\examples{
\dontrun{
generate boxplots:
GCPs_2017_ref <- st_read("I:/misc/GCPs/2017-06/GCPs_ALS_2017_ref.gpkg")
aa_create_boxplots(gcp_data = GCPs_2017_ref, export = T, filename = "Difference_GCPs_ALS_2017",
                   output_path = "I:/misc/GCPs/2017-06")
calculate metrics:
metrics <- aa_metrics(GCPs_2017_ref, export = T, filename = "Metrics_ALS_2017", output_path = "I:/misc/GCPs/2017-06")
metrics
}
Convert data to long format for easier processing
Compute accuracy metrics using yardstick
Export table to csv
Print the final table
}
