% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_classify_ground.R
\name{catalog_classify_ground}
\alias{catalog_classify_ground}
\title{Classify ground points}
\usage{
catalog_classify_ground(
  lascatalog,
  algorithm = lidR::csf(),
  last_returns = T,
  output_path,
  filename_convention = "{ORIGINALFILENAME}",
  parallel = FALSE,
  n_cores = 2
)
}
\arguments{
\item{lascatalog}{object of class \code{lascatalog}.}

\item{algorithm}{an algorithm used for spatial interpolation of the point cloud data, uses the ones available
via \code{\link[lidR:classify]{lidR::classify_ground()}} Defaults to cloth simulation function \code{\link[lidR:gnd_csf]{lidR::csf()}}.}

\item{last_returns}{logical of length 1. Should the computation only be based on last returns? Defaults to TRUE}

\item{output_path}{character path to the folder where the new files should be exported to}

\item{filename_convention}{character defining the filenames of the generated laz files following lidR basics. Defaults
to the original filename}

\item{parallel}{logical of length 1. Should the computation be split over several cores? Defaults to FALSE.}

\item{n_cores}{numeric of length 1. If \code{parall = TRUE}, on how many cores should the computations be run on?
Defaults to the value registered in \code{options("cores")[[1]]}, or, if this is not available, to \verb{parallel::detectCores())}.}
}
\value{
lascatalog
}
\description{
This function classifies ground points in ALS point clouds. It essentially is the same as \link[lidR:classify]{lidR::classify_ground}
but has pre-defined options optimized for the BFNP ALS data.
}
\examples{
ctg <- readALSLAScatalog("/path/to/lazfiles")
ctg_classified <- catalog_normalize(ctg, algorithm = lidR::csf(), last_returns = T,
output_path = "D:/6_pointclouds_classified", filename_convention = "{ORIGINALFILENAME}", parallel = F, n_cores = 1)
apply options to lascatalog
plan parallel processing
classify lascatalog:
}
