% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_clip_polygons.R
\name{catalog_clip_polygons}
\alias{catalog_clip_polygons}
\title{Calculate statistics on lascatalog files}
\usage{
catalog_clip_polygons(
  lascatalog,
  input_epsg,
  output_path,
  filename_convention = "{ID}",
  polygons,
  parallel = F,
  n_cores = 2
)
}
\arguments{
\item{lascatalog}{object of class \code{lascatalog}}

\item{input_epsg}{character EPSG-code of coordinate system}

\item{output_path}{character path to folder where the newly generated laz-files will be exported to}

\item{filename_convention}{character identifying the attribute in the polygon data that should be used to name the
output files}

\item{polygons}{object of class \verb{spatial polygons} featuring an attribute with the IDs or names of the polygons}

\item{parallel}{logical of length 1. Should the computation be split over several cores? Defaults to FALSE.}

\item{n_cores}{numeric of length 1. If \code{parall = TRUE}, on how many cores should the computations be run on?
Defaults to the value registered in \code{options("cores")[[1]]}, or, if this is not available, to \verb{parallel::detectCores())}.}
}
\value{
lascatalog
}
\description{
his function generates multiple output *.laz files based on input polygons (one file per polygon).
The files are named after an attribute in the input polygon dataset.
}
\examples{
ctg <- readALSLAScatalog("/path/to/lazfiles")
catalog_statistics(ctg, TRUE, 3)
check projection if needed:
apply options to lascatalog
plan parallel processing
run clipping:
}
