% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_dtm.R
\name{catalog_dtm}
\alias{catalog_dtm}
\title{DTM creation}
\usage{
catalog_dtm(
  lascatalog,
  resolution = 1,
  algorithm = tin(),
  output_path,
  filename_convention,
  mosaic_result = F,
  mosaic_name = NULL,
  parallel = FALSE,
  n_cores = 2
)
}
\arguments{
\item{lascatalog}{object of class \code{lascatalog}. Needs to have classified ground points.}

\item{resolution}{spatial resolution of the resulting dataset in meters. Corresponds to the projection and coordinate
system of the input dataset. Defaults to 1.}

\item{algorithm}{an algorithm used for spatial interpolation of the point cloud data, uses the ones available
via \code{\link[lidR:rasterize]{lidR::rasterize_terrain()}}. Defaults to \code{tin()}.}

\item{output_path}{character path to the folder where the resulting tif files should be exported to}

\item{filename_convention}{character defining the filenames of the generated tif files following lidR basics. Defaults
to the original filename}

\item{mosaic_result}{logical of length 1. If \code{TRUE}, there will be a mosaiced tif file of the full dataset exported to
the \code{output_path} folder in addition to the tiles.}

\item{mosaic_name}{character representing the filename of the mosaic tif file}

\item{parallel}{logical of length 1. Should the computation be split over several cores? Defaults to FALSE.}

\item{n_cores}{numeric of length 1. If \code{parall = TRUE}, on how many cores should the computations be run on?
Defaults to the value registered in \code{options("cores")[[1]]}, or, if this is not available, to \verb{parallel::detectCores())}.}
}
\value{
SpatRaster list
}
\description{
creates DTMs from a lascatalog only using ground points. The dtms are exported to disc in tiles and
optional, the results can be mosaiced to a full-scale dataset.
}
\examples{
\dontrun{
ctg <- readALSLAScatalog("/path/to/lazfiles")
ctg_dtm <- catalog_dtm(ctg, resolution = 1, algorithm = tin(), output_path = "I:/02_dtms/",
filename_convention = "{ORIGINALFILENAME}", mosaic_result = T, mosaic_name = "DTM_mosaic", parallel = F, n_cores = 1)
}
apply options to lascatalog
function to compute the dtm:
plan parallel processing
apply function to lascatalog:
mosaic the resulting dataset if wanted:
list of all generated tif files:
read in all rasters into a spatial raster collection:
mosaic the files:
export to disk:
}
