% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_statistics.R
\name{catalog_statistics}
\alias{catalog_statistics}
\title{function takes all files in a lascatalog and returns a table with statistics on point density, covex area covered,
extemt, etc.
set catalog options:
function to get point density of las data:
calculate point density:
calculate area covered by convex hull (not bbox):
extract tile name from file (for merging information to other data later):
extract extent of full tile:
make data frame with necessary information:
plan parallel processing
apply function to catalog:
merge the results to a single data frame:}
\usage{
catalog_statistics(lascatalog, parallel = F, n_cores = 2)
}
\description{
function takes all files in a lascatalog and returns a table with statistics on point density, covex area covered,
extemt, etc.
set catalog options:
function to get point density of las data:
calculate point density:
calculate area covered by convex hull (not bbox):
extract tile name from file (for merging information to other data later):
extract extent of full tile:
make data frame with necessary information:
plan parallel processing
apply function to catalog:
merge the results to a single data frame:
}
