% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_subset.R
\name{catalog_subset}
\alias{catalog_subset}
\title{subset lascatalogs}
\usage{
catalog_subset(lascatalog, aoi, output_path, overwrite_files = F)
}
\arguments{
\item{lascatalog}{object of class \code{lascatalog}.}

\item{aoi}{object of class \verb{spatial polygons}}

\item{output_path}{character path to the folder where the new files should be exported to}

\item{overwrite_files}{logical of length 1. Should already existing files be overwritten or be ignored? Defaults to FALSE.}
}
\value{
lascatalog
}
\description{
This function subsets \code{lascatalog} objects using spatial polygon(s) ("aoi, area of interest") and
returns a new \code{lascatalog} object with all tiles overlapping said polygon(s).
In contrast to \code{\link[lidR:catalog_subset]{lidR::catalog_intersect()}} it also exports this new, subsetted catalog to a specified directory very
efficiently without loading the tiles into memory by simply copying the files directly.
This is perfect for when you want to share only sections of the full data with others to work on.
}
\examples{
ctg <- readALSLAScatalog("/path/to/lazfiles")
area_of_interest <- st_read("/path/to/polygons")
ctg_subset <- catalog_subset(ctg, aoi = area_of_interest, output_path = "D:/catalog_subset", overwrite_files = F)
Copy the files that are part of the subset to the new directory;
}
