% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_to_polygons.R
\name{catalog_to_polygons}
\alias{catalog_to_polygons}
\title{Generate footprint polygons of lascatalog objects}
\usage{
catalog_to_polygons(lascatalog)
}
\arguments{
\item{lascatalog}{Object of class \code{lascatalog}}
}
\value{
sf polygon
}
\description{
Takes all files in a lascatalog and returns a polygon file where each polygon represents the bounding box of the respective las / laz file.
In Addition the resulting data is cleaned and unnecessary columns are removed. Data can be exported as geopackage or shapefile afterwards
}
\examples{
ctg <- readALSLAScatalog("/path/to/lazfiles")
catalog_to_polygons(ctg)
convert lascatalog to polygons
convert filename to tile name:
remove unecessary attributes:
rename number of points:
restructure the columns:
}
