% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_gcps.R
\name{merge_GCPs}
\alias{merge_GCPs}
\title{Merge ground control points of 2 datasets for further comparison}
\usage{
merge_GCPs(
  gcp_data_test,
  gcp_data_ref,
  export = F,
  filename = NULL,
  output_path = NULL
)
}
\arguments{
\item{gcp_data_test}{object of class \verb{spatial points} as generated by \verb{bfnpALSprocessor::prepare_gcps())}.
This is the dataset the accurracy should be checked for}

\item{gcp_data_ref}{object of class \verb{spatial points} as generated by \verb{bfnpALSprocessor::prepare_gcps())}.
This is the reference dataset (i.e. the ALS 2019-2020 dataset)}

\item{export}{logical of length 1. Should the result be exported to a geopackage file? Defaults to FALSE.}

\item{filename}{character, optional, name of the resulting geopackage file}

\item{output_path}{character, optional, path to folder where the newly generated gpkg file will be exported to}
}
\value{
sf points
}
\description{
This function takes two outputs of prepare_gcps() and merges them for comparison.
}
\examples{
\dontrun{
merge the GCPs with the reference GCPs:
GCPs_ref <- st_read("I:/misc/GCPs/2019-2020/GCPs_ALS_2019.gpkg")
GCPs_2017 <- st_read("I:/misc/GCPs/2017-06/GCPs_ALS_2017.gpkg")
GCPs_2017_ref <- merge_GCPs(GCPs_2017, GCPs_ref, export = T, filename = "GCPs_ALS_2017_ref", output_path = "I:/misc/GCPs/2017-06")
}
calculate indices describing the nearest feature:
export to disk:
}
