% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_gcps.R
\name{prepare_GCPs}
\alias{prepare_GCPs}
\title{Prepare ground control points derived in cloud compare for further analysis}
\usage{
prepare_GCPs(input_path, output_path, crs = "EPSG:25832", filename, polygons)
}
\arguments{
\item{input_path}{character path to the folder where one or several txt files are stored in the format exported by
the cloud compare point picking tool.}

\item{output_path}{character path to folder where the newly generated gpkg file will be exported to}

\item{crs}{character path, epsg code of coordinate reference system. Defaults to "EPSG:25832"}

\item{filename}{character, name of the geopackage file}

\item{polygons}{object of class \verb{spatial polygons} featuring an attribute with the names of the polygons; used to group
the output by polygon area.}
}
\value{
sf points
}
\description{
This function takes a list of txt-files generated by the CloudCompare point picking tool and converts them to a single
geopackage layer. The points represent GCPs sampled in the point clouds within the pre-defined AOIs.
}
\examples{
\dontrun{
load AOIs:
AOIs <- st_read("I:/misc/test_areas.gpkg", layer = "AOIs_UTM")
prepare the GCPs as generated using CloudCompare:
GCPs_2017 <- prepare_GCPs(input_path = "I:/misc/GCPs/2017-06", output_path = "I:/misc/GCPs/2017-06",
filename = "GCPs_ALS_2017", polygons = AOIs)
}
list all txt files generated by CloudCompare point picking tool:
process the data:
export to disk:
}
