% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_statistics.R
\name{catalog_statistics}
\alias{catalog_statistics}
\title{Calculate statistics on lascatalog files}
\usage{
catalog_statistics(lascatalog, parallel = F, n_cores = 2, spatial = FALSE)
}
\arguments{
\item{lascatalog}{Object of class \code{lascatalog}}

\item{parallel}{logical of length 1. Should the computation be split over several cores? Defaults to FALSE.}

\item{n_cores}{numeric of length 1. If \code{parall = TRUE}, on how many cores should the computations be run on?
Defaults to the value registered in \code{options("cores")[[1]]}, or, if this is not available, to \verb{parallel::detectCores())}.}

\item{spatial}{logical of length 1. Should the output be a spatial dataset (sf) with polygons representing the
catalog tiles?}
}
\value{
data frame
}
\description{
Takes all files in a lascatalog and returns a table with statistics on point density, covex area covered,
extemt, etc. for every tile
}
\examples{
ctg <- readALSLAScatalog("/path/to/lazfiles")
catalog_statistics(ctg, TRUE, 3)
merge the data:
}
